Update main.py
Started to implement the algorithm that converts the heightmap into separate vertices in a numpy array. I'm not sure if it works yet as I am still figuring out how to convert to an STL and view the file in a file editor, but IF all things go accordingly, it should work! (should is the keyword)