Update main.py
I am starting to work on fixing how the program adjusts resolution. Who would have known that that's my largest challenge in this project! Basically what I plan on doing is finding an average of the pixel intensity, and adding padding along the sides where it is needed to make the photo dimensions easily divisible by the largest physical resolution. I'm still not entirely sure if this will work, but lets see!